move SetResourceLimit to util
